[CK] Use double for scaling values at Device Op API
## Motivation

Fixes #4238. Several Device Op APIs added after the original fix
(composable_kernel PR #557) still use `float` or `AccDataType` for
scaling parameters (alpha, epsilon) instead of `double`. This creates
API inconsistency across the CK codebase where older operations
(softmax, reduce, batchnorm, layernorm) correctly use `double` at
the API boundary, but newer WMMA attention and sparse embedding
operations do not.

## Technical Details

Apply the three-layer type specification from the issue:
1. Device Op API: declare scaling/constant args as `double`
2. Device Op storage: convert via `static_cast` to the internal type
   (float for kernel-bound alpha, AccDataType for epsilon)
3. Kernel interface: unchanged (already at correct precision level)

Files changed:
- device_grouped_query_attention_forward_wmma.hpp: alpha float->double
- device_multi_query_attention_forward_wmma.hpp: alpha float->double
- device_batched_gemm_softmax_gemm_permute_wmma_cshuffle.hpp:
  alpha float->double in RawArg, SelfAttnArg, CrossAttnArg
- device_sparse_embeddings_forward_layernorm.hpp:
  epsilon AccDataType->double
- 6 example callers updated for consistency (float->double, stof->stod)

This follows the same pattern established by device_softmax_impl.hpp
which takes `double alpha, double beta` at the API and stores via
`alpha_ = static_cast<AccDataType>(alpha)`.

## Test Plan

No behavioral change - the internal storage types and kernel interface
types remain the same. The `double` API parameters are narrowed via
explicit `static_cast` in Argument constructors, identical to the
existing pattern in device_softmax_impl.hpp, device_reduce.hpp, and
device_batchnorm_forward.hpp. Existing callers that pass `float` values
benefit from implicit widening conversion (float->double is lossless).

+++ projects/composablekernel/include/ck/tensor_operation/gpu/device/impl/device_grouped_query_attention_forward_wmma.hpp
@@ -544,7 +544,7 @@ struct DeviceGroupedQueryAttentionForward_Wmma
                index_t O,
                index_t G0,
                index_t G1,
-               float alpha,
+               double alpha,
                bool input_permute,
                bool output_permute)
             : p_a_grid_{p_a_grid},
@@ -557,7 +557,7 @@ struct DeviceGroupedQueryAttentionForward_Wmma
               O_{O},
               G0_{G0},
               G1_{G1},
-              alpha_{alpha},
+              alpha_{static_cast<float>(alpha)},
               input_permute_{input_permute},
               output_permute_{output_permute}
         {
@@ -590,7 +590,7 @@ struct DeviceGroupedQueryAttentionForward_Wmma
                              index_t O,
                              index_t G0,
                              index_t G1,
-                             float alpha,
+                             double alpha,
                              bool input_permute,
                              bool output_permute)
     {
