+++ projects/composablekernel/include/ck/tensor_operation/gpu/device/impl/device_multi_query_attention_forward_wmma.hpp
@@ -542,7 +542,7 @@ struct DeviceMultiQueryAttentionForward_Wmma
                index_t O,
                index_t G0,
                index_t G1,
-               float alpha,
+               double alpha,
                bool input_permute,
                bool output_permute)
             : p_a_grid_{p_a_grid},
@@ -555,7 +555,7 @@ struct DeviceMultiQueryAttentionForward_Wmma
               O_{O},
               G0_{G0},
               G1_{G1},
-              alpha_{alpha},
+              alpha_{static_cast<float>(alpha)},
               input_permute_{input_permute},
               output_permute_{output_permute}
         {
@@ -588,7 +588,7 @@ struct DeviceMultiQueryAttentionForward_Wmma
                              index_t O,
                              index_t G0,
                              index_t G1,
-                             float alpha,
+                             double alpha,
                              bool input_permute,
                              bool output_permute)
     {
